backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Image processing and computer vision
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# Machine Learning
ultralytics==8.0.196
torch>=2.0.0
torchvision>=0.15.0

# OpenAI API
openai==1.3.5

# Data processing
pandas==2.1.4
python-dateutil==2.8.2

# XML parsing for Goodreads API
lxml==4.9.3

# File handling
aiofiles==23.2.1

# JSON handling
orjson==3.9.10

# Validation
pydantic==2.5.0

# Logging
loguru==0.7.2

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

